+++ ccmpred/objfun/pll/cext/__init__.py
@@ -3,12 +3,12 @@ import numpy.ctypeslib as npct
 import ctypes
 import os.path
 
-array_1d_float = npct.ndpointer(dtype=np.dtype('float32'), ndim=1, flags='CONTIGUOUS')
+array_1d_float = npct.ndpointer(dtype=np.dtype('float64'), ndim=1, flags='CONTIGUOUS')
 array_2d_char = npct.ndpointer(dtype=np.dtype('uint8'), ndim=2, flags='CONTIGUOUS')
 
 libpll = npct.load_library('libpll', os.path.join(os.path.dirname(__file__), '_build'))
 
-libpll.evaluate_pll.restype = ctypes.c_float
+libpll.evaluate_pll.restype = ctypes.c_double
 libpll.evaluate_pll.argtypes = [
     array_1d_float,    # *x
     array_1d_float,    # *g
@@ -18,8 +18,8 @@ libpll.evaluate_pll.argtypes = [
     array_2d_char,      # *msa
     ctypes.c_uint32,    # ncol
     ctypes.c_uint32,    # nrow
-    ctypes.c_float,     # lambda_single
-    ctypes.c_float,     # lambda_pair
+    ctypes.c_double,     # lambda_single
+    ctypes.c_double,     # lambda_pair
 ]
